+++ lib/ruby_lsp/ruby_lsp_yard/addon.rb
@@ -4,12 +4,14 @@ require "ruby-lsp"
 require "ruby_lsp/addon"
 require "uri"
 
+require_relative "../../ruby_lsp_yard/gems"
 require_relative "../../ruby_lsp_yard/indexer"
 require_relative "../../ruby_lsp_yard/inference"
 require_relative "../../ruby_lsp_yard/listeners/completion"
 require_relative "../../ruby_lsp_yard/listeners/definition"
 require_relative "../../ruby_lsp_yard/listeners/hover"
 require_relative "../../ruby_lsp_yard/log"
+require_relative "../../ruby_lsp_yard/rbs"
 require_relative "../../ruby_lsp_yard/settings"
 require_relative "../../ruby_lsp_yard/signature_store"
 require_relative "../../ruby_lsp_yard/version"
@@ -19,13 +21,16 @@ RubyLsp::Addon.depend_on_ruby_lsp!("~> 0.26.0")
 module RubyLsp
   module Yard
     class Addon < ::RubyLsp::Addon
-      attr_reader :settings, :indexer, :signature_store, :inference, :log
+      attr_reader :settings, :indexer, :signature_store, :inference, :log, :rbs_loader, :rbs_source, :gem_cache
 
       def activate(global_state, outgoing_queue)
         @settings = Settings.new(global_state.settings_for_addon(name))
         @log = Log.new(outgoing_queue, level: settings.log_level)
         @indexer = Indexer.for(global_state, log: log)
-        @signature_store = SignatureStore.new(@indexer, log: log)
+        @rbs_loader = build_rbs_loader
+        @rbs_source = Rbs::Source.new(@rbs_loader, log: @log) if @rbs_loader
+        @gem_cache = Gems::Cache.new(locator: Gems::Locator.new, log: @log)
+        @signature_store = SignatureStore.new(@indexer, log: log, rbs: @rbs_source, gem_cache: @gem_cache)
         @inference = Inference::Engine.new(
           adapter: @indexer,
           store: @signature_store,
@@ -40,6 +45,11 @@ module RubyLsp
       end
 
       def deactivate
+        @gem_cache&.flush
+        @rbs_loader&.cancel
+        @rbs_loader = nil
+        @rbs_source = nil
+        @gem_cache = nil
         @inference = nil
         @signature_store = nil
         @indexer = nil
@@ -68,7 +78,6 @@ module RubyLsp
           response_builder,
           node_context,
           dispatcher,
-          store: @signature_store,
           engine: @inference,
           log: @log
         )
@@ -101,6 +110,15 @@ module RubyLsp
           inference: @inference,
           log: @log
         )
+      end
+
+      # NFR-P1: core/stdlib signatures load in the background; features that need them degrade to YARD until ready.
+      def build_rbs_loader
+        return nil unless settings.enabled?(:core_types)
+
+        loader = Rbs::Loader.new(log: @log)
+        loader.start
+        loader
       end
 
       def name