+++ test/ruby_lsp/ruby_lsp_yard/rbs/test_loader.rb
@@ -1,0 +1,53 @@
+# frozen_string_literal: true
+
+require "test_helper"
+require "ruby_lsp_yard/rbs"
+
+module RubyLsp
+  module Yard
+    module Rbs
+      class TestLoader < Minitest::Test
+        def test_loads_core_and_stdlib_synchronously
+          loader = Loader.new(background: false)
+          loader.start
+
+          assert loader.ready?
+          refute_nil loader.environment
+          refute_nil loader.builder
+          assert loader.environment.class_decls.key?(::RBS::TypeName.parse("::String"))
+          assert loader.environment.class_decls.key?(::RBS::TypeName.parse("::Pathname"))
+        end
+
+        def test_loads_in_the_background
+          loader = Loader.new
+          loader.start
+
+          deadline = Process.clock_gettime(Process::CLOCK_MONOTONIC) + 10
+          sleep(0.01) until loader.ready? || Process.clock_gettime(Process::CLOCK_MONOTONIC) > deadline
+
+          assert loader.ready?
+        ensure
+          loader&.cancel
+        end
+
+        def test_skips_libraries_that_do_not_exist
+          loader = Loader.new(background: false)
+          loader.stub(:each_library, ["definitely-not-an-rbs-library"]) do
+            loader.start
+          end
+
+          assert loader.ready?
+          assert loader.environment.class_decls.key?(::RBS::TypeName.parse("::String"))
+        end
+
+        def test_cancel_does_not_publish_a_half_built_environment
+          loader = Loader.new
+          loader.cancel
+          loader.load
+
+          refute loader.ready?
+        end
+      end
+    end
+  end
+end